Don't build the lift-json module for 2.9.3.
Today's magical sbt incantation is *drumroll* SKIP! This will cause the
lift-json project to skip its compile phase if compile is invoked on
dispatch-all. It will still run its update phase, so we have to do some
weird cross version mapping because of the build definition of
unfiltered-json. I think if we can sort out the build issues in
unfiltered-json and have a 2.11 release spun of that, we will be able to
simplify this code.