Instantiate the AsyncHttpClient lazily
This is to fix the resource leak caused by creating a new client in
`configure` and not shutting down the old one.

Instantiating a singleton client was a problem because (by definition of
being a singleton) it might be shared between multiple `Http` instances,
meaning it's hard to know when you can/should shut it down.

Got around the problem by making `Http` hold a config builder and a
`lazy val client`.

It's still possible to leak an `AsyncHttpClient` instance by making
requests and then calling `configure` later, but I think people are less
likely to do that. The normal use case is to configure your client at
application startup and then start making requests.