Update usages of unfiltered to avoid netty compatibility issues
This pulls in 0.8.4 of unfiltered instead of 0.9.0. Although 0.8.4 was
never officially built for Scala 2.12, we spun a custom build of the
parts we needed and stuck it on bintray to resolve an awkward issue with
the underlying netty versions.

We'll back this out as soon as we're able to upgrade to AHC 2.1.