Implement Http.closeAndConfigure
This addresses #53 and other issues in such a way that does not change
underlying behavior in the 0.12.x series. Now, people using Dispatch
0.12.x will get a warning if they attempt to use the configure method
directly.

This has the benefit of not changing the underlying behavior in a minor
release. We will warn using a deprecation warning that this method might
be unsafe, but will permit callers to continue using the Http instance
after invoking it if their application depends on that behavior.

We add a closeAndConfigure method that's the intended "safe" method to
invoke in most cases.