Implement more explicit configuration semantics for Http
This commit forward-ports the API changes from dispatch/reboot#156 into
the 0.13.x release series with a few important changes to boot.

The first, and perhaps biggest, change is that the Http singleton is no
longer an Http executor in it own right. This will no doubt cause some
code breakage that has to be addressed as folks upgrade to 0.13.x. In
order to minimize that code breakage, Http.default provides a default
executor that is always usable for just "doing some HTTP" without
further customization.

This will make explicit the fact that you're invoking a real Thing™ that
has resources that will need to be cleaned up when you're done using
them. This was something that could be ambiguous in earlier versions of
Dispatch because if you happened to invoke `Http()` instead of `Http`
you'd get a new resource pool each time and quickly create problems for
yourself. Two characters should not make that much of a difference in
behavior.

Furthermore, we no longer provide the default builder to the Http case
class by default. It's accessable at Http.defaultClientBuilder if you
wish to use it.

This also adds Http.withConfiguration to facilitate configuring custom
Http instances.