Unwrap execution exceptions if they have a cause
The execution exception itself isn't a terribly useful measure of what actually happened under the
hood. As a result, #111 proposed unwrapping them before finishing the future. I'm still not sure
what I think of this idea, but it was straightforward enough to implement and we have some time
before 0.14.x is declared final.