Document that FunctionHandler isn't thread safe
Clarify that FunctionHandler inherits from a mutable class AsyncCompletionHandler, and therefore each callback should be handled by a fresh instance. 

This PR arose from an actual bug "in the wild" at REA Group in Australia which expended considerable effort to track down. Its an easy mistake to make and Im keen to save the next user from the same confusion.